Add same fix to create_tool_from_function
