+++ releasenotes/notes/fix-component-tool-schema-f27a9c422392b100.yaml
@@ -1,6 +1,6 @@
 ---
 fixes:
   - |
-    Fix ``ComponentTool`` failing to create a tool schema from components with ``Callable`` type parameters
-    (such as ``snapshot_callback``). This enables using ``Agent`` as a ``ComponentTool`` without raising
-    ``SchemaGenerationError``.
+    Fix ``ComponentTool``, ``create_tool_from_function``, and the ``@tool`` decorator failing to create
+    a tool schema when ``Callable`` type parameters are present (such as ``snapshot_callback``).
+    This enables using ``Agent`` as a ``ComponentTool`` without raising ``SchemaGenerationError``.
